% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_background.R
\name{add_background}
\alias{add_background}
\title{Select a background image for a created slide}
\usage{
add_background(background_type)
}
\arguments{
\item{background_type}{string, choose whether the slide should have a title slide appearance "title_slide", or a default content slide appearance "main_slide"}
}
\value{
Returns a backgroud image to be added to your slide
}
\description{
Select a background image for a created slide
}
