% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_formatting.R
\name{format_text}
\alias{format_text}
\title{Format individual lines of text beyond the styles specified in the CSS file}
\usage{
format_text(text, size = 12, colour = "black", weight = "normal")
}
\arguments{
\item{text}{String, text to format according to the specified style}

\item{size}{numeric, font size of formatted text in pixels. Default is 12}

\item{colour}{string, colour of formatted text. Name of a pre-defined HTML colour, or a hex code. Defaults to "black"}

\item{weight}{string, weight of formatted text. Can be "normal", "bold", or "bolder". Default is "normal.}
}
\description{
Formats text on a line-by-line basis
}
