% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_image.R
\name{image}
\alias{image}
\title{Insert an image into a markdown document}
\usage{
image(filepath, alt_text = "", size = NULL, align = "left")
}
\arguments{
\item{filepath}{string, filepath of image to use (must be inside the same project as the Markdown document)}

\item{alt_text}{string, alt text to be used if image cannot be displayed or for screen readers. Defaults to empty}

\item{size}{integer, size of image as a percentage of the original. Default as NULL, image will display as full size}

\item{align}{string, alignment of image on slide. Options are left, middle and right. Defaults to left.}
}
\description{
Insert an image into a markdown document
}
