% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{import_images}
\alias{import_images}
\title{Import background images needed to create slidepacks into your project}
\usage{
import_images(folder_name)
}
\arguments{
\item{folder_name}{Specify the path to import the images into}
}
\description{
Imports the background slide images
Copies the images into a folder in your project directory
}
