% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{import_js_libs}
\alias{import_js_libs}
\title{Import required javascript library into your project}
\usage{
import_js_libs(folder_name)
}
\arguments{
\item{folder_name}{Specify the path to import the library into}
}
\description{
Imports the required JS library
}
