% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic-slides.R
\name{section_slide}
\alias{section_slide}
\title{Create a new section divider slide with header and divider background}
\usage{
section_slide(title = "")
}
\arguments{
\item{title}{String, contents of title field in slide. Defaults to an empty string.}
}
\description{
Create a new section divider slide with header and divider background
}
