% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{setup_project}
\alias{setup_project}
\title{Import required background images and javascript libraries into your project. This only needs to be done once for any project.}
\usage{
setup_project(folder_name = getwd())
}
\arguments{
\item{folder_name}{Specify the path to import the images into (default is the current working directory)}
}
\description{
Import required background images and javascript libraries into your project. This only needs to be done once for any project.
}
