% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidepack_table.R
\name{table}
\alias{table}
\title{Create a css-formatted HTML table to include in your slidepack}
\usage{
table(data, ...)
}
\arguments{
\item{data}{dataframe of data to be displayed in the table}

\item{...}{other arguments to pass to the knitr::kable table}
}
\description{
Create a css-formatted HTML table to include in your slidepack
}
\examples{
table(head(iris))
}
