% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic-slides.R
\name{title_slide}
\alias{title_slide}
\title{Create a new basic title slide containing a title, footer and text content}
\usage{
title_slide(title = "", footer = "", content = "")
}
\arguments{
\item{title}{String, contents of title field in slide. Defaults to an empty string.}

\item{footer}{String, contents of footer field of slide. Defaults to an empty string.}

\item{content}{String, text content of slide. Defaults to an empty string.}
}
\description{
Create a new basic title slide containing a title, footer and text content
}
