% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_functions.R
\name{uneven_cols}
\alias{uneven_cols}
\title{Create a column structure for a slide with a large right column (75\% width) and small left column (25\% width)}
\usage{
uneven_cols(left, right)
}
\arguments{
\item{left}{string, text content to appear in the smaller left column; multiple string objects can be passed as a list.}

\item{right}{any R studio object (chart, table, text) to appear in the larger right column; multiple objects can be passed as a list.}
}
\description{
Create a column structure for a slide with a large right column (75\% width) and small left column (25\% width)
}
\examples{
uneven_cols("This is text", plot(mtcars))
}
