% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_functions.R
\name{even_cols}
\alias{even_cols}
\title{Create a column structure for a slide with equal right and left columns}
\usage{
even_cols(left, right)
}
\arguments{
\item{left}{content to appear in the left column}

\item{right}{content to appear in the right column}
}
\description{
Create a column structure for a slide with equal right and left columns
}
\examples{
even_cols("This is text", plot(mtcars))
}
